COLLADA support part 1.
git-svn-id: svn://0x263a.de/tekuum_src@344 3ae3452d-76a7-4814-9b33-b51e13292e42

+++ libs/irrxml/src/irrXML.h
@@ -376,7 +376,7 @@ namespace io
 	//! defines the utf-32 type. 
 	/** Not using wchar_t for this because 
 	wchar_t has 16 bit on windows and 32 bit on other operating systems. */
-	typedef unsigned long char32;
+	typedef unsigned int char32;
 
 	//! A UTF-8 or ASCII character xml parser.
 	/** This means that all character data will be returned in 8 bit ASCII or UTF-8 by this parser.
